Rename Input (back) to TextInput, File to FileInput
