+++ packages/core/components/field/index.mdx
@@ -6,7 +6,7 @@ route: /components/field
 
 import { Playground, PropsTable } from 'docz'
 import Field from './'
-import Input from '../input'
+import TextInput from '../textInput'
 import Checkbox from '../checkbox'
 import Textarea from '../textarea'
 import { Direction } from "../../common/direction";
@@ -27,7 +27,7 @@ For a group of radio buttons, use [ToggleField](/components/toggleField) instead
 
 <Playground>
     <Field name="Name of your third grade teacher" mb={6}>
-        <Input placeholder="Mrs. Kennedy" />
+        <TextInput placeholder="Mrs. Kennedy" />
     </Field>
 
     <Field
@@ -37,7 +37,7 @@ For a group of radio buttons, use [ToggleField](/components/toggleField) instead
         required
         mb={6}
     >
-        <Input placeholder="Mrs. Eager" required />
+        <TextInput placeholder="Mrs. Eager" required />
     </Field>
 
     <Field name="Describe your fifth grade experience">
@@ -56,7 +56,7 @@ For a group of radio buttons, use [ToggleField](/components/toggleField) instead
         required
         mb={6}
     >
-        <Input placeholder="Arkham Elementary" />
+        <TextInput placeholder="Arkham Elementary" />
     </Field>
 
     <Field
@@ -65,7 +65,7 @@ For a group of radio buttons, use [ToggleField](/components/toggleField) instead
         nameWidth="24rem"
         mb={6}
     >
-        <Input placeholder="Mrs. Hall" />
+        <TextInput placeholder="Mrs. Hall" />
     </Field>
 
     <Field
@@ -77,7 +77,7 @@ For a group of radio buttons, use [ToggleField](/components/toggleField) instead
         required
         mb={6}
     >
-        <Input placeholder="Mrs. Hall" required />
+        <TextInput placeholder="Mrs. Hall" required />
     </Field>
 
     <Field
