+++ packages/core/components/fileInput/index.mdx
@@ -1,16 +1,16 @@
 ---
-name: File
+name: FileInput
 menu: Components
-route: /components/file
+route: /components/fileInput
 ---
 
 import { Playground, PropsTable } from 'docz'
-import File from './'
+import FileInput from './'
 
-# File
+# FileInput
 
 <Playground>
-    <File />
+    <FileInput />
 </Playground>
 
 
@@ -18,4 +18,4 @@ import File from './'
 
 | Prop | Type | Required | Default | Description |
 |:-----|:-----|:---------|:--------|:------------|
-| Input props | | | | see [Input](/components/input) |
+| TextInput props | | | | see [TextInput](/components/textInput) |
