+++ packages/core/components/textInput/index.mdx
@@ -1,34 +1,34 @@
 ---
-name: Input
+name: TextInput
 menu: Components
-route: /component/input
+route: /component/textInput
 ---
 
 import { Playground, PropsTable } from 'docz'
-import Input from './'
+import TextInput from './'
 
-# Input
+# TextInput
 
-Use `Input` for text-like input fields (`text, password, search, email, number,
+Use `TextInput` for text-like input fields (`text, password, search, email, number,
 url, tel`).
 
 ## Basic usage
 
 <Playground>
-    <Input placeholder="Your favorite movie" mb={4} />
-    <Input placeholder="Your favorite dinner" disabled width="45%" mr={8} />
-    <Input placeholder="Your favorite beverage" required width="45%" />
+    <TextInput placeholder="Your favorite movie" mb={4} />
+    <TextInput placeholder="Your favorite dinner" disabled width="45%" mr={8} />
+    <TextInput placeholder="Your favorite beverage" required width="45%" />
 </Playground>
 
 ## Alternate types
 
 <Playground>
-    <Input type="password" placeholder="Password" mb={4} />
-    <Input type="search" placeholder="Search" mb={4} />
-    <Input type="email" placeholder="Email" mb={4} />
-    <Input type="number" placeholder="Number" mb={4} />
-    <Input type="url" placeholder="URL" mb={4} />
-    <Input type="tel" placeholder="Telephone" mb={4} />
+    <TextInput type="password" placeholder="Password" mb={4} />
+    <TextInput type="search" placeholder="Search" mb={4} />
+    <TextInput type="email" placeholder="Email" mb={4} />
+    <TextInput type="number" placeholder="Number" mb={4} />
+    <TextInput type="url" placeholder="URL" mb={4} />
+    <TextInput type="tel" placeholder="Telephone" mb={4} />
 </Playground>
 
 ### PropTypes
