+++ packages/core/components/textarea/index.mdx
@@ -18,4 +18,4 @@ import Textarea from './'
 
 | Prop | Type | Required | Default | Description |
 |:-----|:-----|:---------|:--------|:------------|
-| Input props | | | | see [Input](/components/input) |
+| TextInput props | | | | see [TextInput](/components/textInput) |
