+++ packages/core/index.components.js
@@ -9,13 +9,13 @@ export { default as Card } from "./components/card";
 export { default as Checkbox } from "./components/checkbox";
 export { default as Divider } from "./components/divider";
 export { default as Field } from "./components/field";
-export { default as File } from "./components/file";
+export { default as FileInput } from "./components/fileInput";
 export { default as GlobalStyle } from "./components/globalStyle";
 export { default as Grouper } from "./components/grouper";
 export { default as Header } from "./components/header";
 export { default as Icon } from "./components/icon";
 export { default as Image } from "./components/image";
-export { default as Input } from "./components/input";
+export { default as TextInput } from "./components/textInput";
 export { default as Label } from "./components/label";
 export { default as NavBar } from "./components/navBar";
 export { default as NavBarSection } from "./components/navBarSection";
@@ -25,5 +25,5 @@ export { default as Radio } from "./components/radio";
 export { default as ToggleField } from "./components/toggleField";
 export { default as Swatch } from "./components/swatch";
 export { default as Text } from "./components/text";
-export { default as TextInput } from "./components/textInput";
+export { default as TextTextInput } from "./components/textTextInput";
 export { default as Textarea } from "./components/textarea";
